move processing methods out into individual files
