+++ src/interpreter/processor/token_processors/break.rb
@@ -1,0 +1,9 @@
+module Interpreter
+  class Processor
+    module TokenProcessors
+      def process_break(_token)
+        ReturnValue.new Token::BREAK
+      end
+    end
+  end
+end
