+++ src/interpreter/processor/token_processors/next.rb
@@ -1,0 +1,9 @@
+module Interpreter
+  class Processor
+    module TokenProcessors
+      def process_next(_token)
+        ReturnValue.new Token::NEXT
+      end
+    end
+  end
+end
