+++ src/interpreter/processor/token_processors/no_op.rb
@@ -1,0 +1,9 @@
+module Interpreter
+  class Processor
+    module TokenProcessors
+      def process_no_op(_token)
+        # pass
+      end
+    end
+  end
+end
