+++ src/interpreter/processor/token_processors/return.rb
@@ -1,0 +1,9 @@
+module Interpreter
+  class Processor
+    module TokenProcessors
+      def process_return(_token)
+        ReturnValue.new resolve_variable! @stack
+      end
+    end
+  end
+end
